AnitaRasa-main/accionesdeploy/requirements.txt
rasa-sdk==3.6.2
python-dotenv
psycopg2-binary==2.9.3
requests==2.32.3
reportlab==4.3.1
beautifulsoup4==4.13.3
google-cloud-vision==3.10.0
sib-api-v3-sdk==7.6.0
sanic==21.12.2
SQLAlchemy==1.4.54
